IG-1325 Moved out of collection.route into respository
