+++ src/resources/collections/collections.route.js
@@ -2,22 +2,18 @@ import express from 'express';
 import { ROLES } from '../user/user.roles';
 import passport from 'passport';
 import { utils } from '../auth';
-// import { UserModel } from '../user/user.model'
 import { Collections } from '../collections/collections.model';
 import { Data } from '../tool/data.model';
 import { MessagesModel } from '../message/message.model';
 import { UserModel } from '../user/user.model';
-import emailGenerator from '../utilities/emailGenerator.util';
 import helper from '../utilities/helper.util';
 import _ from 'lodash';
 import escape from 'escape-html';
-import { getCollectionObjects } from './collections.repository';
+import { getCollectionObjects, sendEmailNotifications, generateCollectionEmailSubject } from './collections.repository';
 
 const inputSanitizer = require('../utilities/inputSanitizer');
 
 const urlValidator = require('../utilities/urlValidator');
-
-const hdrukEmail = `enquiry@healthdatagateway.org`;
 
 const router = express.Router();
 
@@ -257,129 +253,4 @@ async function createMessage(authorId, collections, activeflag, collectionCreato
 			saveMessage();
 		}
 	}
-}
-
-async function sendEmailNotifications(collections, activeflag, collectionCreator, isEdit) {
-	// Generate URL for linking collection in email
-	const collectionLink = process.env.homeURL + '/collection/' + collections.id;
-
-	// Query Db for all admins or authors of the collection
-	var q = UserModel.aggregate([
-		{ $match: { $or: [{ role: 'Admin' }, { id: { $in: collections.authors } }] } },
-		{ $lookup: { from: 'tools', localField: 'id', foreignField: 'id', as: 'tool' } },
-		{
-			$project: {
-				_id: 1,
-				firstname: 1,
-				lastname: 1,
-				email: 1,
-				role: 1,
-				id: 1,
-			},
-		},
-	]);
-
-	// Use the returned array of email recipients to generate and send emails with SendGrid
-	q.exec((err, emailRecipients) => {
-		if (err) {
-			return new Error({ success: false, error: err });
-		} else {
-			let subject;
-			let html;
-
-			emailRecipients.map(emailRecipient => {
-				if (collections.authors.includes(emailRecipient.id)) {
-					collections.authors.map(author => {
-						if (activeflag === 'active' && author === emailRecipient.id) {
-							subject = generateCollectionEmailSubject(
-								'Creator',
-								collections.publicflag,
-								collections.name,
-								author === collectionCreator.id ? true : false,
-								isEdit
-							);
-							html = generateCollectionEmailContent(
-								'Creator',
-								collections.publicflag,
-								collections.name,
-								collectionLink,
-								author === collectionCreator.id ? true : false,
-								isEdit
-							);
-						}
-					});
-				} else if (activeflag === 'active' && emailRecipient.role === 'Admin') {
-					subject = generateCollectionEmailSubject('Admin', collections.publicflag, collections.name, false, isEdit);
-					html = generateCollectionEmailContent('Admin', collections.publicflag, collections.name, collectionLink, false, isEdit);
-				}
-
-				emailGenerator.sendEmail([emailRecipient], `${hdrukEmail}`, subject, html, false);
-			});
-		}
-	});
-}
-
-function generateCollectionEmailSubject(role, publicflag, collectionName, isCreator, isEdit) {
-	let emailSubject;
-
-	if (role !== 'Admin' && isCreator !== true) {
-		if (isEdit === true) {
-			emailSubject = `The ${
-				publicflag === true ? 'public' : 'private'
-			} collection ${collectionName} that you are a collaborator on has been edited`;
-		} else {
-			emailSubject = `You have been added as a collaborator on the ${
-				publicflag === true ? 'public' : 'private'
-			} collection ${collectionName}`;
-		}
-	} else {
-		emailSubject = `${role === 'Admin' ? 'A' : 'Your'} ${
-			publicflag === true ? 'public' : 'private'
-		} collection ${collectionName} has been ${isEdit === true ? 'edited' : 'published'} and is now live`;
-	}
-
-	return emailSubject;
-}
-
-function generateCollectionEmailContent(role, publicflag, collectionName, collectionLink, isCreator, isEdit) {
-	return `<div>
-				<div style="border: 1px solid #d0d3d4; border-radius: 15px; width: 700px; margin: 0 auto;">
-					<table
-					align="center"
-					border="0"
-					cellpadding="0"
-					cellspacing="40"
-					width="700"
-					word-break="break-all"
-					style="font-family: Arial, sans-serif">
-						<thead>
-							<tr>
-								<th style="border: 0; color: #29235c; font-size: 22px; text-align: left;">
-									${generateCollectionEmailSubject(role, publicflag, collectionName, isCreator, isEdit)}
-								</th>
-								</tr>
-								<tr>
-								<th style="border: 0; font-size: 14px; font-weight: normal; color: #333333; text-align: left;">
-									${
-										publicflag === true
-											? `${role === 'Admin' ? 'A' : 'Your'} public collection has been ${
-													isEdit === true ? 'edited on' : 'published to'
-											  } the Gateway. The collection is searchable on the Gateway and can be viewed by all users.`
-											: `${role === 'Admin' ? 'A' : 'Your'} private collection has been ${
-													isEdit === true ? 'edited on' : 'published to'
-											  } the Gateway. Only those who you share the collection link with will be able to view the collection.`
-									}
-								</th>
-							</tr>
-						</thead>
-						<tbody style="overflow-y: auto; overflow-x: hidden;">
-							<tr style="width: 100%; text-align: left;">
-								<td style=" font-size: 14px; color: #3c3c3b; padding: 5px 5px; width: 50%; text-align: left; vertical-align: top;">
-									<a href=${collectionLink}>View Collection</a>
-								</td>
-							</tr>
-						</tbody>
-					</table>
-				</div>
-			</div>`;
 }
